Create r_distro
This code generates a separation radius distribution function based on the observed data from binary 1987. We use lucy deconvolution to achieve this. This code was largely generated by chatGPT so it needs to be looked over to ensure all is running as intended. However, plots generate and normalized correctly resembles that form the paper, so it likely is being implemented correctly.